requirements.txt
Flask>=3.1,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
python-dotenv>=1.0,<2.0
SQLAlchemy>=2.0,<3.0
PyMySQL>=1.1,<2.0
Werkzeug>=3.0,<4.0
openpyxl>=3.1,<4.0
